Move Router into Services namespace
