+++ packages/framework/tests/Feature/RouterTest.php
@@ -10,17 +10,17 @@ use Hyde\Framework\Models\Pages\DocumentationPage;
 use Hyde\Framework\Models\Pages\MarkdownPage;
 use Hyde\Framework\Models\Pages\MarkdownPost;
 use Hyde\Framework\Models\Route;
-use Hyde\Framework\Modules\Routing\Router;
+use Hyde\Framework\Services\Router;
 use Hyde\Testing\TestCase;
 use Illuminate\Support\Collection;
 
 /**
- * @covers \Hyde\Framework\Modules\Routing\Router
+ * @covers \Hyde\Framework\Services\Router
  */
 class RouterTest extends TestCase
 {
     /**
-     * @covers \Hyde\Framework\Modules\Routing\Router::getInstance
+     * @covers \Hyde\Framework\Services\Router::getInstance
      */
     public function test_get_instance_returns_the_router_instance()
     {
@@ -31,8 +31,8 @@ class RouterTest extends TestCase
     /**
      * Test route autodiscovery.
      *
-     * @covers \Hyde\Framework\Modules\Routing\Router::__construct
-     * @covers \Hyde\Framework\Modules\Routing\Router::getRoutes
+     * @covers \Hyde\Framework\Services\Router::__construct
+     * @covers \Hyde\Framework\Services\Router::getRoutes
      */
     public function test_get_routes_returns_discovered_routes()
     {
@@ -47,7 +47,7 @@ class RouterTest extends TestCase
     }
 
     /**
-     * @covers \Hyde\Framework\Modules\Routing\Router::getRoutesForModel
+     * @covers \Hyde\Framework\Services\Router::getRoutesForModel
      */
     public function test_get_routes_for_model_returns_only_routes_for_the_given_model()
     {
@@ -65,9 +65,9 @@ class RouterTest extends TestCase
     /**
      * Test route autodiscovery.
      *
-     * @covers \Hyde\Framework\Modules\Routing\Router::discover
-     * @covers \Hyde\Framework\Modules\Routing\Router::discoverRoutes
-     * @covers \Hyde\Framework\Modules\Routing\Router::discoverPageRoutes
+     * @covers \Hyde\Framework\Services\Router::discover
+     * @covers \Hyde\Framework\Services\Router::discoverRoutes
+     * @covers \Hyde\Framework\Services\Router::discoverPageRoutes
      */
     public function test_discover_routes_finds_and_adds_all_pages_to_route_collection()
     {
